+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Persist settings as YAML file in appropriate OS config directory
 - Settings will be saved periodically when using the app
 - Add `save` button to settings window
+- Add ability to change docker socket location in settings